requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2023.2
sqlparse==0.4.3
tzdata==2023.2
